Rename classes and test to project name
